// backend/src/interfaces/location.interface.ts
export interface ILocation {
  id?: string | number;
  locationAccount: string;
  businessName: string;
  dbaName: string;
  streetAddress: string;
  city: string;
  zipCode: string;
  locationDescription: string;
  mailAddress: string;
  mailCity: string;
  mailZipCode: string;
  naics: string;
  primaryNaics: string;
  councilDistrict: string;
  startDate: string;
  endDate: string;
  location: string;
}